requirements.txt
paramiko~=2.11.0
tqdm~=4.64.0
requests~=2.28.1
PyYAML~=6.0